refactor: move AST walker from lint::walker to ast_walker (BT-961)
The shared AST expression walker lived in `lint/walker.rs` but was
already consumed by `semantic_analysis::validators`, making the import
path `crate::lint::walker` architecturally misleading.

Move to a new top-level `ast_walker` module so both lint passes and
semantic validators import from a neutral location that reflects its
cross-cutting role.
